+++ android/src/main/kotlin/in/techbyvishesh/app_utils/AppUtilsPlugin.kt
@@ -2,25 +2,32 @@ package `in`.techbyvishesh.app_utils
 
 
 import android.content.Context
+import android.content.Intent
+import android.widget.Toast
 import androidx.annotation.NonNull
 
 import io.flutter.embedding.engine.plugins.FlutterPlugin
+import io.flutter.embedding.engine.plugins.activity.ActivityAware
+import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
 
 /** AppUtilsPlugin */
-class AppUtilsPlugin: FlutterPlugin, MethodCallHandler {
+class AppUtilsPlugin: FlutterPlugin, MethodCallHandler, ActivityAware {
 
   private lateinit var channel : MethodChannel
   private lateinit var context: Context
+  private lateinit var intent: Intent
 
   override fun onAttachedToEngine(@NonNull flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
     channel = MethodChannel(flutterPluginBinding.binaryMessenger, "app_utils")
     channel.setMethodCallHandler(this)
     context = flutterPluginBinding.applicationContext
   }
+
+  
 
   override fun onMethodCall(@NonNull call: MethodCall, @NonNull result: Result) {
     when(call.method){
@@ -43,6 +50,10 @@ class AppUtilsPlugin: FlutterPlugin, MethodCallHandler {
         context.getAppInfo(result)
       }
 
+      Methods.READ_LAUNCHED_DATA->{
+         context.readLaunchedData(result,intent)
+      }
+
       else -> result.notImplemented()
     }
   }
@@ -50,4 +61,14 @@ class AppUtilsPlugin: FlutterPlugin, MethodCallHandler {
   override fun onDetachedFromEngine(@NonNull binding: FlutterPlugin.FlutterPluginBinding) {
     channel.setMethodCallHandler(null)
   }
+
+  override fun onAttachedToActivity(binding: ActivityPluginBinding) {
+      intent = binding.activity.intent
+  }
+
+  override fun onDetachedFromActivityForConfigChanges() = Unit
+
+  override fun onReattachedToActivityForConfigChanges(binding: ActivityPluginBinding) = Unit
+
+  override fun onDetachedFromActivity() = Unit
 }